+++ forui/lib/src/widgets/sonner/sonner_style.dart
@@ -5,6 +5,19 @@ import 'package:meta/meta.dart';
 
 part 'sonner_style.style.dart';
 
+/// How the sonner should be expanded.
+enum FSonnerExpandBehavior {
+  /// The sonner is always expanded.
+  always,
+
+  /// The sonner is expanded on hover or pressed.
+  onHoverOrPressed,
+
+  /// The sonner is never expanded.
+  disabled,
+}
+
+
 /// [FSonner]'s style.
 class FSonnerStyle with Diagnosticable, _$FSonnerStyleFunctions {
   /// The maximum number of entries shown per [FSonnerAlignment]. Defaults to to 3.
@@ -14,6 +27,18 @@ class FSonnerStyle with Diagnosticable, _$FSonnerStyleFunctions {
   /// The sonner's padding. Defaults to `EdgeInsets.symmetric(horizontal: 20, vertical: 15)`.
   @override
   final EdgeInsetsGeometry padding;
+
+  /// The sonner's expansion behavior. Defaults to [FSonnerExpandBehavior.onHoverOrPressed].
+  @override
+  final FSonnerExpandBehavior expandBehavior;
+
+  /// The duration to wait after entering the sonner before expanding the toasts. Defaults to 200ms.
+  @override
+  final Duration expandHoverEnterDuration;
+
+  /// The duration to wait after exiting the sonner before collapsing the toasts. Defaults to 300ms.
+  @override
+  final Duration expandHoverExitDuration;
 
   /// The spacing below or above the toasts when they are expanded. Defaults to 16.0.
   @override
@@ -53,6 +78,9 @@ class FSonnerStyle with Diagnosticable, _$FSonnerStyleFunctions {
   const FSonnerStyle({
     this.max = 3,
     this.padding = const EdgeInsets.symmetric(horizontal: 20, vertical: 15),
+    this.expandBehavior = FSonnerExpandBehavior.onHoverOrPressed,
+    this.expandHoverEnterDuration = const Duration(milliseconds: 200),
+    this.expandHoverExitDuration = const Duration(milliseconds: 300),
     this.expandStartSpacing = 16,
     this.expandSpacing = 10,
     this.expandDuration = const Duration(milliseconds: 500),
@@ -67,11 +95,11 @@ class FSonnerStyle with Diagnosticable, _$FSonnerStyleFunctions {
 class FToastStyle with Diagnosticable, _$FToastStyleFunctions {
   /// The toast's entrance & exit animation duration. Defaults to 400ms.
   @override
-  final Duration entranceExitDuration;
+  final Duration enterExitDuration;
 
   /// The toast's entrance animation curve. Defaults to [Curves.easeOutCubic].
   @override
-  final Curve entranceCurve;
+  final Curve enterCurve;
 
   /// The toast's exit animation curve. Defaults to [Curves.easeOutCubic].
   @override
@@ -100,8 +128,8 @@ class FToastStyle with Diagnosticable, _$FToastStyleFunctions {
 
   /// Creates a [FToastStyle].
   const FToastStyle({
-    this.entranceExitDuration = const Duration(milliseconds: 400),
-    this.entranceCurve = Curves.easeOutCubic,
+    this.enterExitDuration = const Duration(milliseconds: 400),
+    this.enterCurve = Curves.easeOutCubic,
     this.exitCurve = Curves.easeOutCubic,
     this.entranceExitOpacity = 0.0,
     this.transitionDuration = const Duration(milliseconds: 400),