+++ forui/lib/src/widgets/sonner/toaster.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/foundation.dart';
+import 'package:flutter/material.dart';
 import 'package:flutter/widgets.dart';
 import 'package:forui/forui.dart';
 import 'package:forui/src/widgets/sonner/animated_toaster.dart';
@@ -33,6 +34,8 @@ class Toaster extends StatefulWidget {
 class _ToasterState extends State<Toaster> with SingleTickerProviderStateMixin {
   late final AnimationController _controller;
   late final Animation<double> _expand;
+  int _monotonic = 0;
+  bool _hovered = false;
 
   @override
   void initState() {
@@ -40,14 +43,30 @@ class _ToasterState extends State<Toaster> with SingleTickerProviderStateMixin {
     _controller = AnimationController(vsync: this, duration: widget.style.expandDuration)
       ..addListener(() => setState(() {}));
     _expand = _controller.drive(CurveTween(curve: widget.style.expandCurve));
+
+    if (widget.style.expandBehavior == FSonnerExpandBehavior.always) {
+      _controller.value = 1;
+    }
   }
 
   @override
-  void didUpdateWidget(Toaster oldWidget) {
-    super.didUpdateWidget(oldWidget);
-    if (oldWidget.style != widget.style) {
+  void didUpdateWidget(Toaster old) {
+    super.didUpdateWidget(old);
+    if (old.style != widget.style) {
       _controller.duration = widget.style.expandDuration;
       _expand = _controller.drive(CurveTween(curve: widget.style.expandCurve));
+    }
+
+    if (widget.style.expandBehavior != old.style.expandBehavior) {
+      if (widget.style.expandBehavior == FSonnerExpandBehavior.always) {
+        _controller.value = 1;
+      } else if (widget.style.expandBehavior == FSonnerExpandBehavior.disabled) {
+        _controller.value = 0;
+      }
+    }
+
+    if (widget.entries.isEmpty && widget.style.expandBehavior == FSonnerExpandBehavior.onHoverOrPressed) {
+      _controller.value = 0;
     }
   }
 
@@ -58,10 +77,8 @@ class _ToasterState extends State<Toaster> with SingleTickerProviderStateMixin {
   }
 
   @override
-  Widget build(BuildContext context) => MouseRegion(
-    onEnter: (_) => _controller.forward(),
-    onExit: (_) => _controller.reverse(),
-    child: AnimatedToaster(
+  Widget build(BuildContext context) {
+    Widget toaster = AnimatedToaster(
       style: widget.style,
       alignTransform: widget.alignTransform,
       expand: _expand.value,
@@ -80,6 +97,43 @@ class _ToasterState extends State<Toaster> with SingleTickerProviderStateMixin {
             child: entry.builder(context, entry),
           ),
       ],
-    ),
-  );
+    );
+
+    if (widget.style.expandBehavior == FSonnerExpandBehavior.onHoverOrPressed) {
+      toaster = MouseRegion(
+        onEnter: (_) => _enter(),
+        onExit: (_) => _exit(),
+        child: GestureDetector(
+          onTap: () {
+            if (!_hovered) {
+              _controller.isForwardOrCompleted ? _controller.reverse() : _controller.forward();
+            }
+          },
+          child: toaster,
+        ),
+      );
+    }
+
+    return toaster;
+  }
+
+  Future<void> _enter() async {
+    final fencingToken = ++_monotonic;
+    _hovered = true;
+    await Future.delayed(widget.style.expandHoverEnterDuration);
+
+    if (fencingToken == _monotonic && mounted) {
+      await _controller.forward();
+    }
+  }
+
+  Future<void> _exit() async {
+    final count = ++_monotonic;
+    _hovered = false;
+    await Future.delayed(widget.style.expandHoverExitDuration);
+
+    if (count == _monotonic && mounted) {
+      await _controller.reverse();
+    }
+  }
 }