+++ forui/lib/src/widgets/toast/toaster_layer.dart
@@ -1,7 +1,0 @@
-
-
-import 'package:meta/meta.dart';
-
-/// How the toast should be expanded.
-@internal
-enum ExpandMode { alwaysExpanded, expandOnHover, expandOnTap, disabled }